Extract encode and decode methods to object core
