+++ packages/core/echo/echo-schema/src/automerge/automerge-array.ts
@@ -145,9 +145,9 @@ export class AutomergeArray<T> implements Array<T> {
 
       // TODO(mykola): Do not allow direct access to doc in array.
 
-      const encodedItems = items.map((value) => this._object!._encode(value));
+      const encodedItems = items.map((value) => this._object!._core.encode(value));
 
-      this._object._change((doc) => {
+      this._object._core.change((doc) => {
         let parent = doc;
         for (const key of fullPath.slice(0, -1)) {
           parent = parent[key];
@@ -320,8 +320,8 @@ export class AutomergeArray<T> implements Array<T> {
       const fullPath = [...this._object._core.mountPath, ...this._path!];
 
       // TODO(mykola): Do not allow direct access to doc in array.
-      const encodedItems = items.map((value) => this._object!._encode(value));
-      this._object._change((doc) => {
+      const encodedItems = items.map((value) => this._object!._core.encode(value));
+      this._object._core.change((doc) => {
         let parent = doc;
         for (const key of fullPath.slice(0, -1)) {
           parent = parent[key];
@@ -418,10 +418,10 @@ export class AutomergeArray<T> implements Array<T> {
 
     const fullPath = [...this._object._core.mountPath, ...this._path!];
 
-    const encodedValue = this._object!._encode(value);
+    const encodedValue = this._object!._core.encode(value);
 
     // TODO(mykola): Do not allow direct access to doc in array.
-    this._object._change((doc) => {
+    this._object._core.change((doc) => {
       let parent = doc;
       for (const key of fullPath.slice(0, -1)) {
         parent = parent[key];
